docker_env/requirements.txt
attackcti
plotly
requests
pyyaml
fpdf
Pillow
matplotlib
seaborn
pandas
openhunt
kaleido
